+++ ios/Plugin/Plugin.swift
@@ -15,8 +15,10 @@ public class BranchDeepLinks: CAPPlugin {
                 name: NSNotification.Name.BranchDidStartSession,
                 object: nil
         )
+        
+        Branch.getInstance().registerPluginName("Capacitor", version: "5.0.0")
     }
-    
+
     @objc public func setBranchService(branchService: Any) {
         self.branchService = branchService as! BranchService
     }
@@ -52,10 +54,10 @@ public class BranchDeepLinks: CAPPlugin {
         let linkProperties = getLinkProperties(analytics: analytics, properties: properties)
         let params = NSMutableDictionary();
         params.addEntries(from: linkProperties.controlParams)
-        
+
         branchService.generateShortUrl(params: params as? [AnyHashable : Any] ?? [:], linkProperties: linkProperties) { (url, error) in
             if (error == nil) {
-                call.success([
+                call.resolve([
                     "url": url ?? ""
                 ])
             } else {
@@ -69,15 +71,15 @@ public class BranchDeepLinks: CAPPlugin {
         let properties = call.getObject("properties") ?? [:]
         let shareText = call.getString("shareText", "Share Link")
         let linkProperties = getLinkProperties(analytics: analytics, properties: properties)
-        
+
         let params = NSMutableDictionary();
         params.addEntries(from: linkProperties.controlParams)
-        
+
         let buo = BranchUniversalObject.init()
         DispatchQueue.main.async {
-            buo.showShareSheet(with: linkProperties, andShareText: shareText, from: self.bridge.viewController, completion: nil)
+            buo.showShareSheet(with: linkProperties, andShareText: shareText, from: self.bridge?.viewController, completion: nil)
         }
-        
+
         call.success()
     }
 
@@ -94,17 +96,17 @@ public class BranchDeepLinks: CAPPlugin {
             BranchStandardEvent.rate,
             BranchStandardEvent.search,
             BranchStandardEvent.share,
-            BranchStandardEvent.spendCredits,
             BranchStandardEvent.unlockAchievement,
             BranchStandardEvent.viewCart,
             BranchStandardEvent.viewItem,
             BranchStandardEvent.viewItems,
         ]
 
-        call.success([
+        call.resolve([
             "branch_standard_events": standardEvents
         ])
     }
+
 
     @objc func sendBranchEvent(_ call: CAPPluginCall) {
         guard let eventName = call.options["eventName"] as? String else {
@@ -113,7 +115,7 @@ public class BranchDeepLinks: CAPPlugin {
         }
         let metaData = call.getObject("metaData") ?? [:]
         let event = BranchEvent.customEvent(withName: eventName)
-        
+
         for (key, value) in metaData {
             if key == "transactionID" {
                 event.transactionID = value as? String
@@ -139,18 +141,27 @@ public class BranchDeepLinks: CAPPlugin {
                 event.alias = value as? String
             } else if key == "customData" {
                 event.customData = value as? [String : String] ?? [:]
+            } else if key == "contentMetadata" {
+                if let items = value as? NSMutableArray {
+                    var contentItems: [BranchUniversalObject] = []
+                    
+                    items.forEach { (item) in
+                        let item = item as? [String:Any] ?? [:]
+                        contentItems.append(getContentItemObject(item: item))
+                    }
+                    event.contentItems = contentItems
+                }
             }
         }
-        
-        event.logEvent()
 
+        event.logEvent()
         call.success()
     }
 
     @objc func disableTracking(_ call: CAPPluginCall) {
         let isEnabled = call.getBool("isEnabled") ?? false
         branchService.disableTracking(isEnabled: isEnabled) { (enabled) in
-            call.success([
+            call.resolve([
                 "is_enabled": enabled
             ])
         }
@@ -159,7 +170,7 @@ public class BranchDeepLinks: CAPPlugin {
     @objc func setIdentity(_ call: CAPPluginCall) {
         let newIdentity = call.getString("newIdentity") ?? ""
         branchService.setIdentity(newIdentity: newIdentity) { (referringParams, error) in
-            call.success([
+            call.resolve([
                 "referringParams": referringParams ?? [:]
             ])
         }
@@ -168,7 +179,7 @@ public class BranchDeepLinks: CAPPlugin {
     @objc func logout(_ call: CAPPluginCall) {
         branchService.logout() {(loggedOut, error) in
             if (error == nil) {
-                call.success([
+                call.resolve([
                     "logged_out": loggedOut as Any
                 ])
             } else {
@@ -176,10 +187,10 @@ public class BranchDeepLinks: CAPPlugin {
             }
         }
     }
-    
+
     func getLinkProperties(analytics: [String: Any], properties: [String: Any]) -> BranchLinkProperties {
         let linkProperties = BranchLinkProperties.init()
-        
+
         for (key, value) in analytics {
             if key == "alias" {
                 linkProperties.alias = value as? String
@@ -197,11 +208,99 @@ public class BranchDeepLinks: CAPPlugin {
                 linkProperties.tags = value as? [Any]
             }
         }
-        
+
         for (key, value) in properties {
             linkProperties.addControlParam(key, withValue: value as? String)
         }
-        
+
         return linkProperties
+    }
+    
+    func getContentItemObject(item: [String:Any]) -> BranchUniversalObject {
+        let object = BranchUniversalObject.init()
+
+        for (key, value) in item {
+            
+            switch key {
+            case "productName":
+                object.contentMetadata.productName = value as? String
+            case "productBrand":
+                object.contentMetadata.productBrand = value as? String
+            case "sku":
+                object.contentMetadata.sku = value as? String
+            case "price":
+                object.contentMetadata.price = NSDecimalNumber(decimal: (value as? NSNumber ?? 0).decimalValue)
+            case "currency":
+                object.contentMetadata.currency = (value as? String).map { BNCCurrency(rawValue: $0) }
+            case "quantity":
+                object.contentMetadata.quantity = Double.init(truncating: value as? NSNumber ?? 0)
+            default:
+                object.contentMetadata.customMetadata[key] = value
+            }
+        }
+        
+        return object
+    }
+
+    @objc func getBranchQRCode(_ call: CAPPluginCall) {
+        let analytics = call.getObject("analytics") ?? [:]
+        let properties = call.getObject("properties") ?? [:]
+        let linkProperties = getLinkProperties(analytics: analytics, properties: properties)
+
+        let buo = BranchUniversalObject.init()
+        
+        let qrCodeSettingsMap = call.getObject("settings") ?? [:]
+        let qrCode = BranchQRCode()
+        
+        if let codeColor = qrCodeSettingsMap["codeColor"] as? String {
+            qrCode.codeColor = colorWithHexString(hexString: codeColor)
+        }
+        if let backgroundColor = qrCodeSettingsMap["backgroundColor"] as? String {
+            qrCode.backgroundColor = colorWithHexString(hexString: backgroundColor)
+        }
+        if let centerLogo = qrCodeSettingsMap["centerLogo"] as? String{
+            qrCode.centerLogo = centerLogo
+        }
+        if let margin = qrCodeSettingsMap["margin"] as? NSNumber {
+            qrCode.margin = margin
+        }
+        if let width = qrCodeSettingsMap["width"] as? NSNumber {
+            qrCode.width = width
+        }
+        if let imageFormat = qrCodeSettingsMap["imageFormat"] as? String {
+            if imageFormat == "JPEG" {
+                qrCode.imageFormat = .JPEG
+            } else if imageFormat == "PNG" {
+                qrCode.imageFormat = .PNG
+            }
+        }
+        
+        branchService.getBranchQRCode(branchQRCode: qrCode, buo: buo, linkProperties: linkProperties) { qrCodeString, error in
+            if (error == nil) {
+                call.resolve([
+                    "qrCode": qrCodeString ?? ""
+                ])
+            } else {
+                call.reject(error?.localizedDescription ?? "Error generating QR code")
+            }
+        }
+    }
+    
+    func colorWithHexString(hexString: String, alpha:CGFloat = 1.0) -> UIColor {
+        let hexint = Int(self.intFromHexString(hexStr: hexString))
+        let red = CGFloat((hexint & 0xff0000) >> 16) / 255.0
+        let green = CGFloat((hexint & 0xff00) >> 8) / 255.0
+        let blue = CGFloat((hexint & 0xff) >> 0) / 255.0
+        
+        let color = UIColor(red: red, green: green, blue: blue, alpha: alpha)
+        return color
+    }
+    
+    func intFromHexString(hexStr: String) -> UInt32 {
+        var hexInt: UInt32 = 0
+        let scanner: Scanner = Scanner(string: hexStr)
+        scanner.charactersToBeSkipped = CharacterSet(charactersIn: "#")
+        hexInt = UInt32(bitPattern: scanner.scanInt32(representation: .hexadecimal) ?? 0)
+        return hexInt
     }
 }