+++ CHANGELOG.md
@@ -1,5 +1,33 @@
 Branch Capacitor SDK change log
 
+- 5.0.0
+
+  - Updates the core capacitor plugin to 4.0.1
+  - Android minSdk bumped to 22
+  - iOS platform bumped to 13
+  - Android Branch SDK updated to 5.2.3 which targets 32 as well.
+  - Android Branch SDK dependency type changed from `implementation` to `api` to keep references consistent and integration simpler
+
+- 4.2.0
+
+  - Update Android SDK to 5.2.0 https://github.com/BranchMetrics/android-branch-deep-linking-attribution/releases/tag/5.2.0
+  - Update iOS SDK to 1.43.1 https://github.com/BranchMetrics/ios-branch-deep-linking-attribution/releases/tag/1.43.1
+  - Added getBranchQRCode() method to generate Branch QR codes.
+
+- 4.1.1
+
+  - Update Android SDK to 5.0.15 https://github.com/BranchMetrics/android-branch-deep-linking-attribution/releases/tag/5.0.15
+  - Update iOS SDK to 1.40.2 https://github.com/BranchMetrics/ios-branch-deep-linking-attribution/releases/tag/1.40.2
+  - Update Capacitor Core to 3.3.2
+
+- 3.0.2
+
+  - Update Android SDK to 5.0.8
+
+- 3.0.1
+
+  - Update Android SDK 5.0.7 and iOS SDK 1.39.2
+
 - 3.0.0
 
   - Update Branch Android SDK to 5.0.3